+++ tasmota/tasmota_template.h
@@ -2359,7 +2359,9 @@ const char kModuleNames[] PROGMEM =
   "Odroid Go|"
 #endif  // USE_ODROID_GO
 //  "ESP32-Solo|"
-//  "WT32-Eth01|"
+#ifdef USE_WT32_ETH01
+  "WT32-Eth01|"
+#endif  // USE_WT32_ETH01
 //  "TTGO Watch|"
 #ifdef USE_M5STACK_CORE2
   "M5Stack Core2|"
@@ -2376,7 +2378,9 @@ const uint8_t kModuleNiceList[] PROGMEM = {
   ODROID_GO,
 #endif  // USE_ODROID_GO
 //  ESP32_SOLO,
-//  WT32_ETH01,
+#ifdef USE_WT32_ETH01
+  WT32_ETH01,
+#endif  // USE_WT32_ETH01
 //  TTGO_WATCH,
 #ifdef USE_M5STACK_CORE2
   M5STACK_CORE2,
@@ -2518,6 +2522,51 @@ const mytmplt kModules[] PROGMEM =
     0                            // Flag
   },
 #endif  // USE_ODROID_GO
+#ifdef USE_WT32_ETH01
+  {                              // WT32_ETH01 - (ESP32)
+    0,                           // 0       (I)O                GPIO0, Ethernet EMAC_REF_CLK
+    AGPIO(GPIO_USER),            // 1       IO     TXD0         GPIO1, U0TXD, CLK_OUT3, EMAC_RXD2
+    AGPIO(GPIO_USER),            // 2       IO                  GPIO2, ADC2_CH2, TOUCH2, RTC_GPIO12, HSPIWP, HS2_DATA0, SD_DATA0
+    AGPIO(GPIO_USER),            // 3       IO     RXD0         GPIO3, U0RXD, CLK_OUT2
+    AGPIO(GPIO_USER),            // 4       IO                  GPIO4, ADC2_CH0, TOUCH0, RTC_GPIO10, HSPIHD, HS2_DATA1, SD_DATA1, EMAC_TX_ER
+    AGPIO(GPIO_USER),            // 5       IO                  GPIO5, RXD Led green
+                                 // 6       IO                  GPIO6, Flash CLK
+                                 // 7       IO                  GPIO7, Flash D0
+                                 // 8       IO                  GPIO8, Flash D1
+    0,                           // 9       IO                  GPIO9, Flash D2, U1RXD
+    0,                           // 10      IO                  GPIO10, Flash D3, U1TXD
+                                 // 11      IO                  GPIO11, Flash CMD
+    AGPIO(GPIO_USER),            // 12      (I)O                GPIO12, ADC2_CH5, TOUCH5, RTC_GPIO15, MTDI, HSPIQ, HS2_DATA2, SD_DATA2, EMAC_TXD3       (If driven High, flash voltage (VDD_SDIO) is 1.8V not default 3.3V. Has internal pull-down, so unconnected = Low = 3.3V. May prevent flashing and/or booting if 3.3V flash is connected and pulled high. See ESP32 datasheet for more details.)
+    0,                           // 13      IO                  GPIO13, Ethernet EMAC_RX_ER
+    AGPIO(GPIO_USER),            // 14      IO                  GPIO14, ADC2_CH6, TOUCH6, RTC_GPIO16, MTMS, HSPICLK, HS2_CLK, SD_CLK, EMAC_TXD2
+    AGPIO(GPIO_USER),            // 15      (I)O                GPIO15, ADC2_CH3, TOUCH3, MTDO, HSPICS0, RTC_GPIO13, HS2_CMD, SD_CMD, EMAC_RXD3         (If driven Low, silences boot messages from normal boot. Has internal pull-up, so unconnected = High = normal output.)
+    AGPIO(GPIO_OUTPUT_HI),       // 16      IO                  GPIO16, Ethernet OSC_ENA
+    AGPIO(GPIO_LEDLNK_INV),      // 17      IO                  GPIO17, Network link led (green)
+    AGPIO(GPIO_ETH_PHY_MDIO),    // 18      IO                  GPIO18, Ethernet MDIO
+    0,                           // 19      IO                  GPIO19, Ethernet TXD0
+    0,                           // 20
+    0,                           // 21      IO                  GPIO21, Ethernet EMAC_TX_EN
+    0,                           // 22      IO      LED         GPIO22, Ethernet EMAC_TXD1
+    AGPIO(GPIO_ETH_PHY_MDC),     // 23      IO                  GPIO23, Ethernet MDC
+    0,                           // 24
+    0,                           // 25      IO                  GPIO25, Ethernet EMAC_RXD0
+    0,                           // 26      IO                  GPIO26, Ethernet EMAC_RXD1
+    0,                           // 27      IO                  GPIO27, Ethernet EMAC_RX_DV
+    0,                           // 28
+    0,                           // 29
+    0,                           // 30
+    0,                           // 31
+    AGPIO(GPIO_USER),            // 32      IO                  GPIO32, CFG
+    AGPIO(GPIO_USER),            // 33      IO                  GPIO33, 485_EN
+    0,                           // 34      I   NO PULLUP       GPIO34, ADC1_CH6, RTC_GPIO4
+    AGPIO(GPIO_USER),            // 35      I   NO PULLUP       GPIO35, ADC1_CH7, RTC_GPIO5
+    AGPIO(GPIO_USER),            // 36      I   NO PULLUP       GPIO36, SENSOR_VP, ADC_H, ADC1_CH0, RTC_GPIO0
+    0,                           // 37          NO PULLUP
+    0,                           // 38          NO PULLUP
+    AGPIO(GPIO_USER),            // 39      I   NO PULLUP       GPIO39, SENSOR_VN, ADC1_CH3, ADC_H, RTC_GPIO3
+    0                            // Flag
+  },
+#endif  // USE_WT32_ETH01
 #ifdef USE_M5STACK_CORE2
   {                              // M5STACK CORE2 - (ESP32)
     AGPIO(GPIO_USER),            // 0       (I)O                GPIO0, SPKR_LRCK